pyproject.toml
[tool.poetry]
name = "fiap-hackathon"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.103.2"
uvicorn = "^0.18.3"
pyfunctional = "^1.4.3"
sqlalchemy = "^2.0.8"
sqlalchemy-utils = "^0.40.0"
alembic = "^1.10.2"
psycopg2-binary = "^2.9.1"
dependency-injector = "^4.41.0"
charset-normalizer = "2.1.0"
pydantic-settings = "^2.0.3"
aio-pika = "^9.4.0"
marshmallow-dataclass = "^8.6.0"
pytest-cov = "^4.1.0"


[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.1"
mypy = "^1.1.1"
isort = "^5.12.0"
black = "^23.3.0"
sqlalchemy-stubs = "^0.4"
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-env = "^0.8.2"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
testfixtures = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.Python
    | \.git
    | \.hg
    | \.mypy_cache
    | \__pycache__
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''

[tool.mypy]
python_version = "3.11"
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
allow_untyped_globals = false
allow_redefinition = false
plugins = "sqlmypy"
strict = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.infrastructure.persistence.migrations.*"
disallow_untyped_defs = false
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "factory.*",
    "sqlalchemy_utils.*",
    "uvicorn.*",
    "functional",
    "gunicorn.*",
    "motor.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
combine_as_imports = true
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
multi_line_output = 3
line_length = 120
known_local_folder = ["app", "tests"]
skip = "migrations,tests"

[tool.coverage.run]
omit = [
    "app/infrastructure/rabbit_mq/*",
    "app/domain/shared/message_bus/*",
    "app/presentation/helpers/*",
    "app/shared/*",
    "tests/*"
]